Added a little wall lamp scene, added interpolation for FOV changes, and added some new meshes
Edgar-
I am tired so yuh ill keep this short
added light, a fence mesh and a box crate mesh. the box physics mesh kinda buggy when the player collision collides with it sooo yeahh idk
Added the light so when the interaction system is being tested we can test it out with maybe having a button to turn the light on and off.

Added FOV interpolation, and fixed the crawl not setting speed back to 3. I think thats alll the changes for this rn