MAP MAP + Trees
Edga - OKKKK so  made a rough draft of a map, just basic geometry, plus made a new tree scene that we can just copy and paste. This is a super small and shallow update but just something to get the ball rolling again fr fr Also we still gotta fix the item switching and stuff.